% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/words.R
\docType{data}
\name{alpha_words}
\alias{alpha_words}
\title{tibble of alpha english words}
\format{a grouped tibble with columns
- `word`: the word
- `size`: the number of characters
- `first`: the first letter of the word}
\source{
[dwyl/english-words](https://github.com/dwyl/english-words) on github
}
\usage{
alpha_words
}
\description{
tibble of alpha english words
}
\keyword{datasets}
