% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/words.R
\docType{data}
\name{data_english}
\alias{data_english}
\title{tibble of anagrams in english}
\format{tibble with columns
- word: word
- size: number of letters in the word
- first: first letter of the word
- anagrams: list of anagrams of the word (list column)
- n_anagrams: number of anagrams}
\usage{
data_english
}
\description{
tibble of anagrams in english
}
\keyword{datasets}
