% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/words.R
\docType{data}
\name{english}
\alias{english}
\title{vector of english words}
\format{character vector of english words in lexicographic order}
\source{
[dwyl/english-words](https://github.com/dwyl/english-words) on github
}
\usage{
english
}
\description{
vector of english words
}
\examples{
sample(english, 10)
}
\keyword{datasets}
