% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrabble_points_and_frequencies.R
\docType{data}
\name{scrabble_points_and_frequencies}
\alias{scrabble_points_and_frequencies}
\title{tibble of points and frequencies of letters (a.k.a. tiles) as they appear in English-language edition of the Scrabble game}
\format{tibble with columns
- letter: letters from A to Z plus Blank/Wild tiles
- points: points (or values) of each letter in the game, from 0 to 10
- freq: frequency of each letter in the game, total of 100}
\source{
[Scrabble Tile Distribution and Point Values](https://www.thespruce.com/scrabble-tile-distribution-and-point-values-412402)
}
\usage{
scrabble_points_and_frequencies
}
\description{
tibble of points and frequencies of letters (a.k.a. tiles) as they appear in English-language edition of the Scrabble game
}
\keyword{datasets}
