% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scrabble_points_and_frequencies.R
\docType{data}
\name{scrabble_points_and_frequencies}
\alias{scrabble_points_and_frequencies}
\title{tibble of points and frequencies of letters (a.k.a. tiles) as they appear in English-language edition of the Scrabble game}
\format{tibble with columns
- letter: letter A to Z plus Blank/Wild tiles
- points: points  letter of the word
- freq: frequency of the letter in the Scrabble game}
\source{
[Scrabble Tile Distribution and Point Value](https://www.thespruce.com/scrabble-tile-distribution-and-point-values-412402)
}
\usage{
scrabble_points_and_frequencies
}
\description{
tibble of points and frequencies of letters (a.k.a. tiles) as they appear in English-language edition of the Scrabble game
}
\keyword{datasets}
